pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sony-automator-controls"
version = "1.0.2"
description = "Sony Automator Controls - TCP to HTTP command bridge for Cuez Automator"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "BlueElliott", email = "elliott.ramdass10@gmail.com"}
]
keywords = ["sony", "automator", "tcp", "http", "control", "broadcast", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
]

build = [
    "pyinstaller>=6.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/BlueElliott/Elliotts-Sony-Automator-Controls"
Repository = "https://github.com/BlueElliott/Elliotts-Sony-Automator-Controls"
Issues = "https://github.com/BlueElliott/Elliotts-Sony-Automator-Controls/issues"
Documentation = "https://github.com/BlueElliott/Elliotts-Sony-Automator-Controls#readme"

[project.scripts]
sony-automator-controls = "sony_automator_controls.__main__:main"

[tool.setuptools]
packages = ["sony_automator_controls"]

[tool.setuptools.package-data]
sony_automator_controls = ["*.json"]
"*" = ["static/*"]

# ==================== BLACK ====================
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

# ==================== RUFF ====================
[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

# ==================== BANDIT ====================
[tool.bandit]
exclude_dirs = [".venv", "venv", "build", "dist"]
skips = ["B101"]
